# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ShopIA API",
    "version": "1.0.0",
    "description": "Documentation API ShopIA Dashboard"
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "paths": {
...
    "/api/boutiques/livraison/entreprises": {
      "get": {
        "summary": "Suivi des entreprises de livraison",
        "description": "Tableau de suivi détaillé par entreprise avec toutes les métriques financières",
        "tags": [
          "Livraison"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Recherche par nom d'entreprise"
          },
          {
            "in": "query",
            "name": "filter",
            "schema": {
              "type": "string",
              "enum": [
                "all_partenaires",
                "actifs",
                "inactifs"
              ],
              "default": "all_partenaires"
            }
          },
          {
            "in": "query",
            "name": "period",
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "week",
                "month",
                "year"
              ],
              "default": "month"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entreprises récupérées avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "entreprise": {
                                "type": "string"
                              },
                              "partenaires": {
                                "type": "integer"
                              },
                              "cmd_mois": {
                                "type": "integer"
                              },
                              "stock_gere": {
                                "type": "integer"
                              },
                              "montant_collecte": {
                                "type": "number"
                              },
                              "commission": {
                                "type": "number"
                              },
                              "reverse": {
                                "type": "number"
                              },
                              "depenses": {
                                "type": "number"
                              },
                              "taux_livraison": {
                                "type": "number"
                              },
                              "taux_livraison_statut": {
                                "type": "string",
                                "enum": [
                                  "bon",
                                  "moyen",
                                  "critique"
                                ]
                              },
                              "incidents": {
                                "type": "integer"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/boutiques/livraison/depenses": {
      "get": {
        "summary": "Dépenses enregistrées des entreprises de livraison",
        "description": "Retourne le détail des dépenses par catégorie pour chaque entreprise",
        "tags": [
          "Livraison"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "period",
            "schema": {
              "type": "string",
              "enum": [
                "today",
                "week",
                "month",
                "year"
              ],
              "default": "month"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Dépenses récupérées avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "entreprise": {
                            "type": "string"
                          },
                          "carburant": {
                            "type": "number"
                          },
                          "communication": {
                            "type": "number"
                          },
                          "frais_divers": {
                            "type": "number"
                          },
                          "total_depenses": {
                            "type": "number"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}